apg/templates/aiohttp/framework/{{cookiecutter.app_name}}/requirements.txt
aiohttp==3.1.1
gunicorn==19.7.1
uvloop==0.9.1
ipdb==0.11
webargs==5.1.0
click==6.7
pprintpp==0.3.0
apispec==0.39.0
python-dotenv==0.8.2
yarl==1.2.4
{% if cookiecutter.use_database == 'y' %}
# database requirements
alembic==1.0.7
asyncpg==0.18.2
gino==0.8.1
psycopg2-binary==2.7.7
SQLAlchemy==1.2.17
SQLAlchemy-serializer==1.1
pgcli==2.0.2
{% endif %}
# test requirements
pytest==3.6.3
pytest-cov==2.5.1
pytest-aiohttp==0.3.0
pylama==7.6.6
pylint==2.2.2
isort==4.3.4
mccabe==0.6.1
